de forulles zitten er in. je moet yii migrate/redo uitvoegen de laste migratie is ook aangepast

+++ console/migrations/m160412_081647_changeNorm.php
@@ -6,13 +6,36 @@ class m160412_081647_changeNorm extends Migration
 {
     public function safeUp()
     {
-        $this->renameColumn('norm_category', 'score', 'max');
-        $this->addColumn('norm_category', 'formule', 'blob');
+        if (in_array('score', $this->getModelColumns())) {
+            $this->renameColumn('norm_category', 'score', 'max');
+        }
+        if (!in_array('formule', $this->getModelColumns())) {
+            $this->addColumn('norm_category', 'formule', 'text');
+        }
+
+        if (!in_array('default', $this->getModelColumns())) {
+            $this->addColumn('norm_category', 'default', 'int(2)');
+        }
+    }
+
+    private function getModelColumns()
+    {
+        $model = new \common\models\NormCategory();
+        return $model->getTableSchema()->getColumnNames();
     }
 
     public function safeDown()
     {
-        $this->renameColumn('norm_category', 'max', 'score');
-        $this->dropColumn('norm_category', 'formule');
+        if (in_array('max', $this->getModelColumns())) {
+            $this->renameColumn('norm_category', 'max', 'score');
+        }
+
+        if (in_array('formule', $this->getModelColumns())) {
+            $this->dropColumn('norm_category', 'formule');
+        }
+        if (in_array('default', $this->getModelColumns())) {
+            $this->dropColumn('norm_category', 'default');
+        }
     }
+
 }
